Move misc helpers to common
